Dynamic form changes with validation
