+++ src/app/app-routing.module.ts
@@ -1,7 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { DynamicFormComponent } from './components/dynamic-form/dynamic-form.component';
 
-const routes: Routes = [];
+const routes: Routes = [{
+  path: '',
+  component: DynamicFormComponent
+}];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
