+++ src/app/app.component.scss
@@ -16,51 +16,5 @@
         margin: 0 20px;
       }
     }
-    form {
-      margin-top: 20px; 
-      padding-bottom: 20px; 
-
-      height: calc(100% - 140px);
-      formly-form {
-        height: calc(100% - 10px);
-        overflow: auto;
-        position: relative;
-        display: block;
-        padding: 0px 20px;
-      }
-    }
-
-    .footer-container {
-      height: 70px;
-      background: lightgray;
-      padding: 10px 20px;
-      margin-left: -5px;
-    }
-
-    .form-group {
-      width: calc(50% - 10px);
-      margin-bottom: 10px;
-    }
-  
-    button {
-      margin: 5px;
-    }
-  
-    ::ng-deep mat-form-field,
-    ::ng-deep formly-field-mat-datepicker {
-      margin: 5px;
-    }
-  
-    ::ng-deep .mat-mdc-text-field-wrapper {
-      background-color: transparent !important;
-    }
-  
-    ::ng-deep .mat-mdc-form-field-infix {
-      margin-top: 5px;
-    }
-  
-    ::ng-deep .mat-mdc-form-field-error {
-      color: red;
-    }
   }
