+++ src/app/components/dynamic-form/dynamic-form.component.html
@@ -1,0 +1,7 @@
+<form [formGroup]="form" (ngSubmit)="onSubmit()">
+    <formly-form [model]="model" [fields]="fields" [form]="form"></formly-form>
+    <div class="col-md-12 footer-container">
+      <button type="button" class="col-md-1 btn btn-outline-primary submit-button" (click)="onCancel()">Cancel</button>
+      <button type="submit" class="col-md-1 btn btn-primary submit-button">Submit</button>
+    </div>
+  </form>
