+++ src/app/components/dynamic-form/dynamic-form.component.scss
@@ -1,0 +1,56 @@
+:host {
+    display: flex;
+    justify-content: left;
+    align-items: left;
+    height: 100vh;
+    position: relative;
+    width: 100%;
+    flex-direction: column;
+    overflow: hidden;
+    ::ng-deep .form-label {
+      display: none;
+    }
+
+    nav {
+      color: white;
+      height: 80px;
+      div {
+        margin: 0 20px;
+      }
+    }
+    form {
+      margin-top: 20px; 
+      padding-bottom: 20px; 
+
+      height: calc(100% - 60px);
+      formly-form {
+        height: calc(100% - 10px);
+        overflow: auto;
+        position: relative;
+        display: block;
+        padding: 0px 20px;
+      }
+    }
+
+    .footer-container {
+      height: 70px;
+      background: lightgray;
+      padding: 10px 20px;
+      margin-left: -5px;
+    }
+
+    .form-group {
+      width: calc(50% - 10px);
+      margin-bottom: 10px;
+    }
+  
+    button {
+      margin: 5px;
+    }
+  
+    ::ng-deep mat-form-field,
+    ::ng-deep formly-field-mat-datepicker {
+      margin: 5px;
+    }
+  }
+
