+++ src/app/components/formly-field-numetic.component.ts
@@ -5,8 +5,8 @@ import { FieldType, FieldTypeConfig } from '@ngx-formly/core';
  selector: 'formly-field-input',
  template: `
   <div class="col-sm-2 form-element form-group">
-    <label class="label" for="field-input-text">{{label}} <span class="label-required" *ngIf="isRequired">(required)</span></label>
-    <input name="field-input-text" type="number" required="true" class="border border-primary form-control" [formControl]="formControl" [formlyAttributes]="field">
+    <label class="label"  [for]="field.key">{{label}} <span class="label-required" *ngIf="isRequired">(required)</span></label>
+    <input [name]="field.key" [type]="field.props.type" class="border border-primary form-control" [formControl]="formControl" [formlyAttributes]="field">
   </div>
  `,
  styles: `
