+++ src/assets/mockData.json
@@ -4,43 +4,66 @@
             "label": "First Name",
             "type": "varchar",
             "key": "fname",
-            "value": "John"
+            "value": "John",
+            "validation": {
+                "required": true,
+                "minLength": 2
+            }
         },
         {
             "label": "Last Name",
             "type": "varchar",
             "key": "lname",
-            "value": "Doss"
+            "value": "Doss",
+            "validation": {
+                "required": true,
+                "minLength": 2
+            }
         },
         {
             "label": "Email",
             "type": "varchar",
             "key": "email",
-            "value": "John.Doss@abc.com"
+            "value": "John.Doss@abc.com",
+            "validation": {
+                "required": true
+            }
         },
         {
             "label": "Weight",
             "type": "double",
             "key": "weight",
-            "value": "35"
+            "value": "35",
+            "validation": {
+                "required": true
+            }
         },
         {
             "label": "Height",
             "type": "double",
             "key": "height",
-            "value": "120"
+            "value": "120",
+            "validation": {
+                "required": true
+            }
         },
         {
             "label": "Gender",
             "type": "varchar",
             "key": "gender",
-            "value": "F"
+            "value": "",
+            "validation": {
+                "required": true
+            }
         },
         {
             "label": "Date",
             "type": "date",
             "key": "date",
-            "value": "2024-03-04"
+            "value": "2024-03-04",
+            "validation": {
+                "required": true
+            }
         }
     ]
 }
